Progress - 15

+++ src/main/java/com/payhere/cafe/entity/User.java
@@ -1,7 +1,7 @@
 package com.payhere.cafe.entity;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Data;
 
 @Entity
@@ -10,7 +10,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "phonenum")
     private String phonenum;
